Change server modules structure
