+++ .idea/artifacts/common_jvm_0_1.xml
@@ -1,6 +1,8 @@
 <component name="ArtifactManager">
   <artifact type="jar" name="common-jvm-0.1">
     <output-path>$PROJECT_DIR$/common/build/libs</output-path>
-    <root id="archive" name="common-jvm-0.1.jar" />
+    <root id="archive" name="common-jvm-0.1.jar">
+      <element id="module-output" name="sd-id.common.jvmMain" />
+    </root>
   </artifact>
 </component>
