+++ .idea/artifacts/frontend_0_1.xml
@@ -1,6 +1,8 @@
 <component name="ArtifactManager">
   <artifact type="jar" name="frontend-0.1">
     <output-path>$PROJECT_DIR$/frontend/build/libs</output-path>
-    <root id="archive" name="frontend-0.1.jar" />
+    <root id="archive" name="frontend-0.1.jar">
+      <element id="module-output" name="sd-id.frontend.main" />
+    </root>
   </artifact>
 </component>
