--- common/src/commonMain/kotlin/Common.kt
@@ -1,7 +1,0 @@
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class User(val name: String)
-
-@Serializable
-data class ServerGreeting(val greeting: String)
