+++ server/application/build.gradle.kts
@@ -1,0 +1,32 @@
+plugins {
+    kotlin("jvm")
+}
+
+tasks.test {
+    useJUnitPlatform()
+}
+
+dependencies {
+    implementation(project(":server:base"))
+    implementation(project(":server:domain"))
+    implementation(project(":server:core"))
+    implementation(project(":server:web"))
+    implementation(libs.ktor.client.cio)
+    implementation(libs.ktor.server.cio)
+    implementation(libs.ktor.server.core.jvm)
+    implementation(libs.ktor.server.host.common.jvm)
+    implementation(libs.ktor.server.status.pages.jvm)
+    implementation(libs.ktor.server.content.negotiation)
+    implementation(libs.ktor.server.cors.jvm)
+    implementation(libs.ktor.server.call.logging.jvm)
+    implementation(libs.ktor.server.websockets)
+    implementation(libs.ktor.serialization.kotlinx.json)
+    implementation(libs.kotlinx.coroutines.core.jvm)
+    implementation(libs.logback.classic)
+    implementation(libs.exposed.core)
+    implementation(libs.exposed.dao)
+    implementation(libs.exposed.jdbc)
+    implementation(libs.postgresql)
+    implementation(libs.koin.ktor)
+    implementation(libs.koin.logger.slf4j)
+}
