+++ server/core/build.gradle.kts
@@ -1,0 +1,12 @@
+plugins {
+    kotlin("jvm")
+}
+
+tasks.test {
+    useJUnitPlatform()
+}
+
+dependencies {
+    implementation(project(":server:domain"))
+    implementation(libs.koin.jvm)
+}
