+++ server/core/src/main/kotlin/com/example/secure/chat/core/CoreModule.kt
@@ -1,0 +1,6 @@
+package com.example.secure.chat.core
+
+import org.koin.dsl.module
+
+val coreModule = module {
+}
