+++ server/domain/build.gradle.kts
@@ -1,0 +1,19 @@
+plugins {
+    kotlin("jvm")
+}
+
+tasks.test {
+    useJUnitPlatform()
+}
+
+dependencies {
+    implementation(project(":server:base"))
+    implementation(libs.kotlinx.coroutines.core.jvm)
+    implementation(libs.logback.classic)
+    implementation(libs.exposed.core)
+    implementation(libs.exposed.dao)
+    implementation(libs.exposed.jdbc)
+    implementation(libs.postgresql)
+    implementation(libs.koin.ktor)
+    implementation(libs.koin.logger.slf4j)
+}
