+++ server/domain/src/main/kotlin/com/example/secure/chat/domain/DbModule.kt
@@ -1,0 +1,6 @@
+package com.example.secure.chat.domain
+
+import org.koin.core.module.Module
+
+val dbModules = listOf<Module>(
+)
