--- server/src/main/kotlin/com/example/auth/server/modules/ApplicationModules.kt
@@ -1,6 +1,0 @@
-package com.example.auth.server.modules
-
-val applicationModules = listOf(
-    dbModule,
-    servicesModule
-)
