--- server/src/main/kotlin/com/example/auth/server/modules/ServicesModule.kt
@@ -1,8 +1,0 @@
-package com.example.auth.server.modules
-
-import org.koin.dsl.module
-import com.example.auth.server.services.UserService
-
-val servicesModule = module {
-    single { UserService() }
-}
