--- server/src/main/kotlin/com/example/auth/server/routing/IndexRouting.kt
@@ -1,11 +1,0 @@
-package com.example.auth.server.routing
-
-import io.ktor.server.http.content.*
-import io.ktor.server.routing.*
-
-fun Routing.indexRouting() {
-    static("/") {
-        resources("/static")
-        resource("/", "static/index.html")
-    }
-}
