--- server/src/main/kotlin/com/example/auth/server/routing/api/ApiRouting.kt
@@ -1,10 +1,0 @@
-package com.example.auth.server.routing.api
-
-import io.ktor.server.routing.*
-import com.example.auth.server.routing.api.users.userRouting
-
-fun Routing.apiRouting() {
-    route("/api") {
-        userRouting()
-    }
-}
