--- server/src/main/kotlin/com/example/auth/server/routing/routingUtils.kt
@@ -1,13 +1,0 @@
-@file:Suppress("unused")
-
-package com.example.auth.server.routing
-
-import io.ktor.http.*
-
-fun <T> Parameters.saveReceive(name: String, map: (String) -> T?): T {
-    return get(name)?.let(map) ?: throw IllegalArgumentException("Invalid $name parameter")
-}
-
-fun Parameters.saveReceive(name: String): String {
-    return saveReceive(name) { it }
-}
