--- server/src/main/kotlin/com/example/auth/server/util/resourceUtils.kt
@@ -1,5 +1,0 @@
-package com.example.auth.server.util
-
-fun Class<*>.resourceBytes(resourcePath: String): ByteArray? {
-    return classLoader.getResourceAsStream(resourcePath).use { it?.readBytes() }
-}
