--- server/src/main/resources/application.conf
@@ -1,36 +1,0 @@
-ktor {
-  environment = production
-  environment = ${?KTOR_ENV}
-
-  application {
-    modules = [com.example.auth.server.ApplicationKt.module]
-  }
-
-  deployment {
-    host = 127.0.0.1
-    port = 8080
-  }
-
-  logger {
-    level = INFO
-    level = ${?KTOR_LOGGER_LEVEL_INFO}
-  }
-}
-
-db {
-  postgres {
-    host = localhost
-    host = ${?DB_POSTGRES_HOST}
-
-    port = 5678
-    port = ${?DB_POSTGRES_PORT}
-
-    username = postgres
-    username = ${?DB_POSTGRES_USERNAME}
-
-    password = postgres
-    password = ${?DB_POSTGRES_PASSWORD}
-
-    name = postgres
-  }
-}
