+++ server/web/src/main/kotlin/com/example/secure/chat/web/WebModule.kt
@@ -1,0 +1,6 @@
+package com.example.secure.chat.web
+
+import org.koin.dsl.module
+
+val webModule = module {
+}
