+++ server/web/src/main/kotlin/com/example/secure/chat/web/routing/api/ApiRouting.kt
@@ -1,0 +1,8 @@
+package com.example.secure.chat.web.routing.api
+
+import io.ktor.server.routing.*
+
+fun Routing.apiRouting() {
+    route("/api") {
+    }
+}
