+++ server/web/src/main/kotlin/com/example/secure/chat/web/util/resourceUtils.kt
@@ -1,0 +1,5 @@
+package com.example.secure.chat.web.util
+
+fun Class<*>.resourceBytes(resourcePath: String): ByteArray? {
+    return classLoader.getResourceAsStream(resourcePath).use { it?.readBytes() }
+}
