+++ settings.gradle.kts
@@ -1,3 +1,11 @@
 rootProject.name = "sd-id"
 
-include("common", "frontend", "server")
+include(
+    "common",
+    "frontend",
+    "server:base",
+    "server:domain",
+    "server:core",
+    "server:web",
+    "server:application",
+)
